refactor: rename files to fallow js conventions
