+++ src/events/interactionCreate.ts
@@ -1,5 +1,5 @@
 import { Interaction } from 'discord.js';
-import handleInteractions from '../lib/interaction-handlers.js';
+import handleInteractions from '../lib/interactionHandler.js';
 import { GoogleGenAI } from '@google/genai';
 
 export const name = 'interactionCreate';
