+++ src/lib/interactionHandler.ts
